Fix span length calculation bugs.
git-svn-id: https://svn.apache.org/repos/asf/pivot/trunk@955970 13f79535-47bb-0310-9956-ffa450edef68

+++ wtk/src/org/apache/pivot/wtk/Span.java
@@ -76,7 +76,7 @@ public final class Span {
      * The absolute value of (<tt>end</tt> minus <tt>start</tt>) + 1.
      */
     public long getLength() {
-        return Math.abs(end - start) + 1;
+        return Math.abs((long)end - (long)start) + 1;
     }
 
     /**
